permutation matrix must be strictly integer to avoid rounding down
If user-supplied permutation matrix is not strictly integer, the
values can be rounded down, for instance, sqrt(3)^2 will be
truncated to 2 when handled as an integer. We were permissive in
other functions and rounded to nearest integer if input were
practically integer, but we are strict in permutation matrices
and request user to handle non-integer data.

(cherry picked from commit dc6d6bd8d736dadc07a01fe33a0f665cf1ec3866)