do not round to integers if data were stored as integers
Most integer data are numeric in R, and they are stored as double
with all decimals as zero. After any operation on such data, the
decimal part may have a non-zero bit and the analysis can fail if
they are treated as being exactly integer. Therefore we round after
checking that data are practically integer (within sqrt(.Machine$double.eps)
of whole number) which is unnecessary if data were exact integers.
However, if the data are also stored as integer, we can be sure that
there is no decimal noise, and we skip the round().