sort species counts only when the number of species is high
Sorting takes time and is useful only with high number of species.
I #define'd a limit of 100 species (which can be changed in
compile time with flag -D SORTLIMIT). This was based on
benchmarking with BCI and BCI multiplied 5-fold in both dims to
250x1125. Probably the sample size also influences the limit
as do data set properties. However, in extreme cases this
seems to be useful.

All these changes (with as.matrix fix in the master branch)
make  4 to 6 -fold speed-up.