check that data were found in model.frame.cca()
removed test against prc(), because new test catches that too.
We cannot reconstruct model frame in prc(), when the formula was
used or when data= was missing (attached data frame or variables
in the .GlobalEnv). Nor do we try harder in those cases (it could
be possible at least for the vars in .GlobalEnv).