try to fix the bothced utils::read.fortran() decimal handling
if the real (F) format for abundance data defines decimals (such as
F5.1), read.fortran will divide the input with the correspoding
power of 10 even when the input had explicit decimal sign (which
violates Fortran standard). We try to live with this, and if the
input has values with more non-zero decimals than defined in the
format, we multiply with the corresponding power of ten to undo
this stupid R decision.

If R Core ever comes to senses, this can be cancelled, but I
think the function is usually correct even if if R is fixed.

(cherry picked from commit d8938de4f50fb4f6c228f8be9dff759c6262c769)