try parallelization of vegdist.c with openMP
does not seem to be too useful: in Macbook Air, dissimilarities
of 500 observations take 0.25 sec without parallelization, 0.1 sec
with four threads. It is a speed-up, but both are fast. When tried
with 12900 observations and 1000 species, time was cut from 640 sec
to 550 sec, or 90 sec. However, still slow.

This version tries to have more balanced scheduling by collapsing
two for loops. R dist() (and distance.c) parallelize only outer loop
with similar effect.