huge (I mean huge) data sets exhausted memory
12900 sample plots (83M+ dissimilarities) exhausted the stack
(if "Illegal instruction: 4" in macOS means that). I assume that
the reason was reduction(+:dist) of length 83M+. That was unnecessary,
because the values are collected in the vector with private index.
There were some other clean-ups, and together all these fixed the
problem, although I assume that fixing the reduction() clause was the
main thing