use natural way of calculationg RsquareAdj for cca: conflicts varpart
varpart adjusts R2 of basic components, and R2-adj for a testable
partial model is the difference of adjusted models. This version
of RsquareAdj.cca directly estimates the R2 of the partial model and
adjusts that, or in other words, adjusts a difference of R2 values
of taking difference of adjusted R2 values (varpart).