get around the limit of 500 characters when deparsing Condition
R function deparse() has a maximum limit of 500 characters for
strings and long Condition() statements in constrained ordination
could be broken into several strings with confusing error message
(could not find function "Condition"). Now we have a kluge to
circumvent this hard limit.

The real solution is to have better and simpler ordiParseFormula
instead of stacking kluge upon a kluge.

The issue was behind stackoverlow query
https://stackoverflow.com/questions/49249816/ordistep-function-from-vegan-error-on-large-rdas