permutest.cca by="terms" failed with socket clusters
With by="terms" the permutest.cca evaluates all contrasts within
compiled code and the number of columns in return matrix depends
on the number of contrasts. Socket clusters in parallel processing
where not adapted to this, but assumed only tests of one term.
Socket clusters are used as the only alternative in Windows, and
optional with pre-set clusters in unix-like systems.

Addresses github issue #276 where this was reported for adonis2.
Obviously also influences by="onedf" in permutest, although this
was not tested.