convex hull failed in plot.betadisper if a group had only one case
https://stackoverflow.com/questions/50267430/ worry
"“Error: Incorrect no.of dimensions” when plotting multivariate data in Vegan"
can be reproduced if one factor level had only one case, and the convex hull
of this one point failed because of dropped dimensions. There are many ways of
fixing this (like checking that there are more than one point for the convex
hull), but this fix takes care the dimensions are not dropped.