resurrect cca-object-tests, but more robustly
This is the same file as the previously removed cca-object-tests
but edited to make it more robust with alternative BLAS/Lapack.

cca-object-tests are compared against reference output, but with
different versions of BLAS/Lapack the output diverged. The two main
reasons were:
1. The sign of eigenvector is undefined. The solutions was to to
drop plain scores output with unstable signs, and use abs() in some
other cases where sign is involved.
2. Last bits vary, and in particular that influenced values that are
almost zero. In some BLAS/Lapcak these could be 5e-33 and in another
5e-34. Now we use zapsmall() to make these practical zeros to real.

This commit is not tested with alternative BLASes: I have only one
in this laptop. Check with OpenBLAS is pending.