relax: test only marginal properties of simulated null models
Earlier we tested that simulated matrices were unchanged, but now
we only check their properties (marginal frequencies, marginal
sums etc) are unchanged. The purpose of the earlier tests was to
see if we can guarantee replication of simulation results across
vegan versions. However, this failed as a general test, because
R could be compiled with different configurations, and quantitative
models using stats::rmultinom() gave different results when R
was built without long double support (used in rmultinom), and this
was reported as a vegan issue in CRAN tests.