do not use constraints or conditions when they have zero columns
We only tested for non-null constraints & conditins in ordConstrained,
but we could have null data in 0-column matrix, and then we actually
ran unconstrained ordination instead of skipping analysis. This happened
in github issue #291, where a marginal anova.cca() gave wrong results
when the model only had one variable: when that one variable was dropped,
we got a zero-column constraint matrix.