vegdist is faster when input is cast to matrix early
Tests with rowSums(x), colSums(x), x<0 etc. as well as decostand(x)
are faster for matrix 'x' than for data.frame 'x', and we need
matrix 'x' finally in the .Call() anyway. So we make this cast in
the beginning instead of doing it as late as possible. vegdist(dune)
is now three times faster (from 0.75 to 0.25 msec).