formula.data.frame changes in R-devel and vegan must adapt
Formerly in R, the formula of model.frame was re-constructed from
the column names of the model.frame, but after R svn revisions
75891, 75911 (changes) and 75915 (export) the saved formula was
used. This broke ordiParseFormula in some special cases. The
ordiParseFormula is a complete mess and cruft upon cruft. This adds
still one layer and circumvents the problem instead of solving.

This will give a NOTE in R-release: R-devel changed
stats:::formula.model.frame and provided a compatibility function
stats::DF2formula which does not exist in R-release and we get
a NOTE on "no visible definition" for a function that will not be
used when it is not visible.