Really dirty kludge to pass R CMD check in R <= 3.6.0 (R-release)
R 3.6.0 changed the behaviour of stats:::formula.data.frame, but
provided a compatibility function DF2formula. We use this function
in R 3.6.0, but R CMD check will issue a NOTE for this function
missing in R version where we do not use it. Now we define this
function when it is never used to keep R happy (and it exists when
it is used and we need to do nothing).

Together with the previous commit, this provides a quick & dirty
fix to github.com/vegandeves/vegan issue #299. ordiParseFormula
is cruft upon cruft, and this adds still one layer. The real
solution is to completely re-write formula interpretation in
constrained ordination, but that will not be quick.