Cleaner (but not more beautiful) to R 3.6.0 formula.data.frame change
For scoping, variable names with functions must be quoted in the
formula (for instance, `poly(A1, 2)`, `log(A1)`) so that the variable
(A1) is later found. stats:::formula.data.frame() did this in
R < 3.6.0, but now it returns unquoted formula and we must be
explicit.

This is a fix to github.com/vegandevs/vegan issue #299, and a bit
cleaner than the previous version. Now we explitly re-construct the
formula with back-quotes when needed, instead of using different
stats functions in R < 3.6.0 (stats:::formula.data.frame) and
R >= 3.6.0 (stats::DF2formula), and need no code dependent on
R version.