start implementing simpler & more orthodox ordiParseFormula
This is triggered by github.com/vegandevs/vegan issue #299. The
purpose is to have much simpler, clearly faster and more robust
version of ordiParseFormula. The work starts from the scratch and
implements first only a part of the original functionality thus
breaking vegan, but we try to gain the original functionality with
time.