squared distances to centroid can be negative in betadisper (issue #306)
Some squared distances to centroid can be negative giving complex
valued distances. We used to take the Mod of these complex value (or
sqrt of abs differences) which replaces these "more similar than
identical" distances with a positive value inflating estimates of
within-group distances. Now we take only the real part (or zero) for
these distances. This is still biased, but less so than the old
practice.