Use R_unif_index() of R instead of our own way of getting index
R_unif_index() was added to the official C API in R 3.6.0, but
it has been unofficially available since R 3.4.x (haven't checked
if x==0 or larger). It was added to the R devel branch on
2017-03-15 by luke (r72356). This means that the new macro would
also succeed in compilation in older R, but only R 3.6.0
has the fixed code.

This will change the test results (but not in oecosimu-tests!),
and there are also cascading effects to other than nestedness
functions.