make vegan dependent on R (>= 3.4.0)
R 3.4.0 is the first that exported R_unif_index in its C API,
although this was advertised as a part of public API only in
R 3.6.0. Although R_unif_index is available retroactively in
pre-3.6.0, the R_unif_index code was improved in R 3.6.0 and
naturally it is wise to upgrade (but that's none of vegan's
business). However, now code is made equal to the corresponding
R code and its base::sample() function. See issue #312.