labdsv 2.0-1 removed rndtaxa: remove x-ref to avoid CRAN problems
labdsv ditches rndtaxa and adds new funciton rndcomm which is not
completely identical, but provides similar alternative as some of
vegan quantitative null model. However, x-ref would be version
dependent, and I completely remove the reference to labdsv function
at the moment.

This is a tiny issue, but still triggered a warning in CRAN, and
I got a message:

Please see the problems shown on
<https://cran.r-project.org/web/checks/check_results_vegan.html>.

Please correct before 2019-08-19 to safely retain your package on CRAN.

Issue #322 in GitHub.