add chord and chi-square distances to vegdist
The initial design idea of vegdist was to separate distance
formulae from data standardization, and these distances could be
calculated as Euclidean distances after decostand standardizations.
It seems that several people have not accepted this design idea, and
assume we do not have those distances. These two popular distances
can now be calculated directly in vegdist which applies decostand()
internally. This is made to serve and protect the people.