adipart, multipart formula add unique id and constant if needed
New behaviour guarantees that there is alpha diveristy for single
SUs and gamma for all SUs pooled. If you do not like this, use
non-formula interface. However, sandwiching levels between these
may add some results, but does not change any of the others. Consider
this is a rare case of user-friendliness.