rename variable 'group' to more appropriate 'score'
ggplot does not automatically map 'group' to group, but this must
be explicit in aesthetics. It really isn't group, and score is a
better name.

Also expand slightly the docs: list the names of scores for each
case and still once emphasize that the results are not suitable
for standard graphics, but only work in ggplot2.