more robust bracketing, and allow meaningles alpha when S==N
We know that the function for uniroot increases at root and used
extendInt="upX", but this gives two long jumps down to negative
values that fail in log(), and therefore we have symmetric
bracketing with shorter leaps which seam to handle more smoothly
pathological estimates.

When all species are singletons (all with count 1), the estimated
alpha are astronomically high, starting from fisherfit(1) ca. 1e8.
Previously we failed with fisherfit(c(1,1,1,1,1,1)) or six singletons
in bracketing, but now these huge values go through smoothly.