R 4.1.0 needs more explicit imports
Error: processing vignette 'FAQ-vegan.Rmd' failed with diagnostics:
The 'markdown' package should be installed and declared as a dependency
of the 'vegan' package (e.g., in the 'Suggests' field of DESCRIPTION),
because the latter contains vignette(s) built with the 'markdown' package.
Please see https://github.com/yihui/knitr/issues/1864 for more information.