dependence tests failed in microbial: change uniroot bracketing
microbial example plotalpha used data with extreme Fisher alphas,
and the bracketing strategy failed: we started with bracket c(1,50),
but some alphas were up to 1035! The new bracketing was adopted in
commit 69b59dc0b4 and this partially reverts that commit. Judging
from the comments on 69b59dc0, there was a reason to switch the
extendInt strategy and there may be new problems looming around.
Certainly something more is needed, but try this first to get
this to CRAN.