2.6-2 is in CRAN: up devel version numbering
The plan is to keep only one working branch, but fork each release
from master with necessary edits, and therefore we need to have
different minor numbers (master for development above latest
release). Previously we had parallel cran-2.5 and master (2.6-0)
branches, but now I don't expect so radical changes in master that
cherry-picking to a CRAN branch would be useful.