change phrasing in help to avoid confusing innocent users
We used to write about convergence problems in help, and this made
people think that the analysis failed if the result could not be
repeated, and made them to switch to other software which did not
study the results and hence did not warn on replication problems.
Now we write about repeating the results and explain that the results
are at least as good as with other software even when they could not
be repeated.