C code for direct calculation Chord & Hellinger distances (unused)
Chord & Hellinger distances can be calculated without prior
standardization (except sqrt for Hellinger). This means collecting
sums of squares and cross product and which can lose precision and
it may be better to use standardization + Euclidean distance, and the
C code is not used currently (but standardization also can lose
precision?).