anova.cca(..., by="margin") for one var set permutation model NA
There is no sense to ask by="margin" anova when there is only one
variable: overall test gives the result. However, if done so, the
name of permutation model could not be set from an empty model, and
was given as NA.

Reported for adonis2 in https://stats.stackexchange.com/questions/575998/
but concerns all anova.cca(..., by="margin") models.