pCCA component did not modify the object passed to CCA
Curiouslu, this really is not a bug, but the results are practically
identical with correct modification of the species data. I think
that the ordination scores u are orthogonal to Conditions (pCCA)
component and therefore its effect need not be removed, and the whole
pCCA step is redundant. I'll keep it, though, as it does not seem
to do any harm, and may protect against something I don't know about.

Found in R CMD check with environment
_R_CHECK_CODETOOLS_PROFILE_="suppressLocalUnused=FALSE"