decorana prints always axis length, used to print only rescaled DCA
The site scores are scaled so that the the heterogeneity is averaged
to unit, whereas rescaling also tries to make the heterogeneity
constant over axis. However, also non-rescaled axis should give
consistent estimate of axis length, although heterogeneity may not
be constant. The scaling is equal to scaling="sites", hill=TRUE in rda.