arrow heads were not drawn to zero-length arrows (with warnings)
Zero-length arrows come with duplicated data or LC scores in
cca/rda. If such duplicated cases are at the end, the arrow to
the end can be of zero length, and then the arrow is indeterminate.
Other cases were handled silenty, but naturally it makes no sense
to draw zero-length lines.