take care that P-values do not decrease along axes in anova.cca
Follows Canoco which uses cumulative maximum for axiswise P-values
in CCA. Usually this has no effect because P-values are increasing.
Most often drops in P-values occurred in last axes when some
previous axes were P=1, but later could be P=0.999 or something.
Now the default is to 'cutoff' these cases.