new dissimilarities need to know about old ones to blend nicely
Previously dissimilarities were calculated from new to fixed
points. We did not know how the new dissimilarities blend among the
old ones, and some of the new points would be blown out of the margins
of the plot. Now we supply also the dissimilarities among fixed
points to allow smooth blending of new dissimilarities and make
dissimmilarities be similarly scaled in getting their ranks in
monotone regression.