qr.X could fail with too many aliased variables: fixes issue #452
model.matrix fails if the full model matrix has more columns than
there are observations, but number of non-aliased parameters was
lower than the number observations. This typically happens with
models with high-order interactions. This was reported for marginal
models in adonis2 (issue #452) but also affects constrained
ordination methods (rda, cca, dbrda) in anova with by="margin" or
by="axis".