base MDSaddpoints on dist2xy (but still accepts dist data)
Purpose is to switch completely to require dist2xy, but keep a
"secret" interface to still work on 'dist', where last rows of
dist are for the new points. This allows using old scripts.

This is checked against previous version and gives identical
results both when verifying that LOO jackknife gives original
coordinates of full matrix, and CV same results in calibration.