fisher diversity works with very species poor communities: issue #588
Lower limit of uniroot was too high to estimate very low diversities:
low number of species with high number of individuals. The current
limits should work with communities of one species in 10,000,000,000
individuals, and still work with all communities of singletons that
fit in R vector. See discussion in issue #588, and comments in
commits  3bce687 and 69b59dc.