warn on overfitted constrained ordination (should this be message?)
I assumed that people will see from the printed output that there
is no residual unconstrained component making permutation tests
etc futile and models useless. This seems not be the case, so issue
a loud warning. (I wonder if this is too intrusive and confusing to
users and we should use less visible message()).