predict.dbrda: can extract wa scores (but not predict with newdata)
The WA scores can be calculated as
   ordiYbar(..., "initial") %*% u %*% diag(1/eigen)
but the initial cannot be found as newdata. If newdata has the
same number of rows as the original data, we could find WA scores
from predict(..., type="lc", newdata=...), but this is hardly
meaningful (and if you think you need it and understand what that
means, you surely are able to code it yourself).