code to return scores with summary during a mercy period
Several packages use summary() to get scores, but this was not
possible after summary dropped scores. To avoid breaking CRAN with
the 2.6-6 release, scores are returned with the object although they
are not printed. If we detect scores related arguments, we also warn
on obsolete practice and tell them to move to scores(). The plan is
to have this only in 2.6-6 release and not in anhy later.