ordiplot: better handling of missing (species) scores
Looking at the code, it seems to assume that empty list is NULL,
or is.null(list()) is TRUE. This may have been the case in 2003
when this code was written, but not since R 2.2-1 from 2005.
Probably the code should be re-written without is.null().

(cherry picked from commit f3c7637cb3b8ed770388d8dd4999a5b89aa09c73)