we should depend on vegan3d to have xref links in man pages
Cross-references \code{\link[vegan3d]{...}} made pkgdown want to
build vegan3d, but rgl was not built with the installation as vegan
did not have dependency on vegan3d and this crashed the build. CRAN
also checks against xref to packages that are not listed as
dependencies (cf commit 172ac2ae56).