getF.c: fix col dimension of Q for Hat matrix: can SEGFAULT!!
Seems to work now, but the function can segfault in BLAS dsyrk
in parallel processing. This seems to be known problem with
BLAS/LAPACK in R (google it!). In my system (Mac M1, Accelerate
BLAS with neural engine) this happend regularly in fork clusters
but socket clusters created with parallel::makeCluster() worked
nicely.