getF.c: remove our own implementation of QR decomposition
We added function do_QR in getF.c since at that time R implementation
dqrdc2 was not in official R API (although it worked) and we wanted
to have plan B if dqrdc2 is unavailable. qrdc2 was added to official
API in R 4.5.0 and we may now safely remove do_QR (and dig it up
from git if needed later).