ordipointlabel: tweak optimization parameters
In particular, starting temperature of simulated annealing looks
influential: higher values allow worse moves above other labels in
the beginning. In ordipointlabel the unit of change is the area of
letter m, and the temperature influences the probability of
accepting change of number of m's.  Number of iterations directly
influences running time, and we adjust this by the number of movable
labels to reduce running time.