optspace does not double-recentre output, but decostand does
Initial centring is not a part of optspace (and never done there),
so the final centring should neither be that the method should return
approximation of the input. However, double-centring is done in
decostand. With this the output of optspace changes, but the output
of decostand ("rclr") and vegdist ("robust.aitchison") remain unchanged.